LayoutTests/imported/w3c/web-platform-tests/content-security-policy/unsafe-hashes/javascript_src_denied_wrong_hash-window_open-expected.txt


FAIL Test that the javascript: src is not allowed to run assert_equals: expected "script-src-elem" but got "script-src"
